Correciones ObraService

+++ src/main/java/co/edu/uniandes/dse/museoartemoderno/services/ObraService.java
@@ -2,12 +2,16 @@ package co.edu.uniandes.dse.museoartemoderno.services;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.Date;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import co.edu.uniandes.dse.museoartemoderno.entities.ObraEntity;
+import co.edu.uniandes.dse.museoartemoderno.entities.MuseoEntity;
+import co.edu.uniandes.dse.museoartemoderno.entities.MovimientoArtisticoEntity;
+import co.edu.uniandes.dse.museoartemoderno.entities.ArtistaEntity;
 import co.edu.uniandes.dse.museoartemoderno.exceptions.EntityNotFoundException;
 import co.edu.uniandes.dse.museoartemoderno.exceptions.IllegalOperationException;
 import co.edu.uniandes.dse.museoartemoderno.repositories.ObraRepository;
@@ -33,6 +37,24 @@ public class ObraService {
 	public ObraEntity createObra(ObraEntity obraEntity) throws EntityNotFoundException, IllegalOperationException {
 		log.info("Inicia el proceso de creación de la Obra");
 		
+		
+		if (obraEntity.getNombre() == null)
+			throw new IllegalOperationException("El Nombre de la Obra ingresado es nulo.");	
+		
+//		if (!obraRepository.findByNombre(obraEntity.getNombre()).isEmpty())
+//			throw new IllegalOperationException("El nombre de la obra ingresado ya existe");
+		
+		if (obraEntity.getArtista() == null)
+			throw new IllegalOperationException("El Artista de la obra ingresada es nulo.");
+		
+		if (obraEntity.getMuseo() == null)
+			throw new IllegalOperationException("El Museo de la obra ingresada es nulo.");	
+		
+		if (obraEntity.getMovimiento() == null)
+			throw new IllegalOperationException("El Movimiento Artístico de la obra ingresada es nulo. ");
+		
+		if ((obraEntity.getFechaPublicacion()).after(new Date()))
+			throw new IllegalOperationException("La fecha de publicación no es válida.");
 	
 		log.info("Termina el proceso de creación de la Obra");
 		return obraRepository.save(obraEntity);
@@ -102,9 +124,22 @@ public class ObraService {
 		log.info("Inicia proceso de borrar la Obra con id: ", obraId);
 		Optional<ObraEntity> obraEntity = obraRepository.findById(obraId);
 		
-		//TODO
 		if (obraEntity.isEmpty())
 			throw new EntityNotFoundException("No se encontró ninguna obra con el ID dado. ");
+	
+		MovimientoArtisticoEntity movimiento = obraEntity.get().getMovimiento();
+
+		if (movimiento != null)
+			throw new IllegalOperationException("No se puede eliminar la obra ya que está asociada a un Movimiento Artístico");
+
+		MuseoEntity museo = obraEntity.get().getMuseo();
+		if (museo != null)
+			throw new IllegalOperationException("No se puede eliminar la obra ya que está asociada a un Museo");
+
+		ArtistaEntity artista = obraEntity.get().getArtista();
+		if (artista != null)
+			throw new IllegalOperationException("No se puede eliminar la obra ya que está asociada a un Artista");
+
 		
 		obraRepository.deleteById(obraId);
 		log.info("Termina proceso de borrar la Obra con id: " + obraId);
